refactor(e2e): Decompose god class subtest_executor.py
Closes #478

- Split 2269-line file into 5 focused modules:
  * parallel_executor.py (766 lines): RateLimitCoordinator, parallel execution, retry logic
  * agent_runner.py (162 lines): Agent execution helpers and result validation
  * judge_runner.py (269 lines): Judge execution, consensus, result validation
  * workspace_setup.py (249 lines): Workspace management and git worktree setup
  * subtest_executor.py (993 lines): Core SubTestExecutor class and orchestration
- Maintained backward-compatible re-exports via __all__ and __getattr__
- All external imports continue to work without modification:
  * scylla/e2e/runner.py imports run_tier_subtests_parallel
  * scylla/e2e/rerun.py imports SubTestExecutor, _commit_test_config
  * Test files import various helper functions
- Updated test patches to reference new module locations
- Added _compute_judge_consensus as SubTestExecutor method for backward compatibility
- All 424 E2E unit tests pass
- Pre-commit hooks pass
