+++ scylla/e2e/agent_runner.py
@@ -1,0 +1,162 @@
+"""Agent execution helpers for E2E testing.
+
+This module handles:
+- Saving and loading agent execution results
+- Creating agent model metadata files
+- Validating agent result integrity
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+import subprocess
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+from scylla.e2e.paths import RESULT_FILE, get_agent_result_file
+
+if TYPE_CHECKING:
+    from scylla.adapters.base import AdapterResult
+
+logger = logging.getLogger(__name__)
+
+
+def _save_agent_result(agent_dir: Path, result: AdapterResult) -> None:
+    """Save agent execution result to agent/result.json.
+
+    Args:
+        agent_dir: Path to agent directory
+        result: AdapterResult from agent execution
+
+    """
+    result_data = {
+        "exit_code": result.exit_code,
+        "stdout": result.stdout,
+        "stderr": result.stderr,
+        "token_stats": result.token_stats.model_dump(),
+        "cost_usd": result.cost_usd,
+        "api_calls": result.api_calls,
+    }
+
+    with open(agent_dir / RESULT_FILE, "w") as f:
+        json.dump(result_data, f, indent=2)
+
+
+def _load_agent_result(agent_dir: Path) -> AdapterResult:
+    """Load agent execution result from agent/result.json.
+
+    Args:
+        agent_dir: Path to agent directory
+
+    Returns:
+        AdapterResult loaded from file
+
+    """
+    from scylla.adapters.base import AdapterResult, AdapterTokenStats
+
+    with open(agent_dir / RESULT_FILE) as f:
+        data = json.load(f)
+
+    token_stats = AdapterTokenStats(**data["token_stats"])
+
+    return AdapterResult(
+        exit_code=data["exit_code"],
+        stdout=data["stdout"],
+        stderr=data["stderr"],
+        token_stats=token_stats,
+        cost_usd=data["cost_usd"],
+        api_calls=data["api_calls"],
+    )
+
+
+def _create_agent_model_md(agent_dir: Path, model: str) -> None:
+    """Create MODEL.md file with agent model and version information.
+
+    Args:
+        agent_dir: Path to agent directory
+        model: Model identifier used for the agent
+
+    """
+    # Try to get claude-code version
+    try:
+        result = subprocess.run(
+            ["claude", "--version"],
+            capture_output=True,
+            text=True,
+            timeout=10,
+        )
+        if result.returncode == 0:
+            claude_code_version = result.stdout.strip()
+        else:
+            claude_code_version = "unknown"
+    except Exception:
+        claude_code_version = "unknown"
+
+    model_info = f"""# Agent Model Information
+
+**Model**: {model}
+**Claude Code Version**: {claude_code_version}
+**Timestamp**: {datetime.now(timezone.utc).isoformat()}
+
+## Configuration
+- Model ID: {model}
+- Output Format: text
+- Permissions: dangerously-skip-permissions (for testing)
+
+## Environment
+- Generated by ProjectScylla E2E Framework
+"""
+
+    agent_dir.mkdir(parents=True, exist_ok=True)
+    (agent_dir / "MODEL.md").write_text(model_info)
+
+
+def _has_valid_agent_result(run_dir: Path) -> bool:
+    """Check if a valid agent result exists for the run.
+
+    A result is considered invalid if:
+    - Result file doesn't exist
+    - JSON is malformed
+    - Required fields are missing
+    - exit_code is -1 AND all token_stats are 0 (incomplete execution)
+
+    Args:
+        run_dir: Path to the run directory
+
+    Returns:
+        True if valid agent result exists, False otherwise
+
+    """
+    result_file = get_agent_result_file(run_dir)
+    if not result_file.exists():
+        return False
+
+    try:
+        data = json.loads(result_file.read_text())
+        # Check all required fields exist
+        required_fields = ["exit_code", "token_stats", "cost_usd"]
+        if not all(field in data for field in required_fields):
+            return False
+
+        # Check for incomplete execution: exit_code=-1 AND all token stats are 0
+        # This indicates the agent threw an exception but created files
+        if data["exit_code"] == -1:
+            token_stats = data["token_stats"]
+            all_tokens_zero = (
+                token_stats.get("input_tokens", 0) == 0
+                and token_stats.get("output_tokens", 0) == 0
+                and token_stats.get("cache_creation_tokens", 0) == 0
+                and token_stats.get("cache_read_tokens", 0) == 0
+            )
+            if all_tokens_zero:
+                logger.warning(
+                    f"Invalid result detected at {run_dir}: "
+                    f"exit_code=-1 with zero token stats (incomplete execution)"
+                )
+                return False
+
+        return True
+    except (json.JSONDecodeError, KeyError, OSError):
+        return False
